Override the "equals" method.
In order to overcome this error we use override equals() and hashCode()
in java.
By using equals() method it compares the objects and the memory location
and checks that the objects are pointing to same memory allocation or
not. hashCode() method must be overridden in every class that overrides
equals() method. hashCode()  invoked  is used to generate an integer
code to that objects.
It is done because the class object does not behave properly on
hash-based functions which are organized in a sequence of buckets. Hash
is used to determine and find where the object is stored and the
position of the bucket. Equals method is used to find the bucket for the
key.